Function rusty_gists::strlen

pub fn strlen(s: impl AsRef<str>) -> usize

Returns length of something that can be converted into a &str

Arguments

s - can be any type that can be converted into a str reference.

Meta

This is an example of how generics can make library function interface more ergonomic for the consumer. The consumer can provide &str, String, or Box<str> and without the rust compiler complaining about these being different types.

Examples

use rusty_gists::strlen;

let my_str = "a";
let my_string = String::from("bc");
let my_boxed_str = String::from("def").into_boxed_str();

assert_eq!(strlen(my_str), 1);
assert_eq!(strlen(my_string), 2);
assert_eq!(strlen(my_boxed_str), 3);

References

std::convert::AsRef